requirements.txt
pyyaml
dicttoxml
mysql-connector-python